[feat/#19] 본인이 작성한 일기만 삭제 가능하도록 구현

+++ src/main/java/com/aiary/aiary/domain/diary/service/DiaryService.java
@@ -7,6 +7,10 @@ import com.aiary.aiary.domain.diary.entity.Diary;
 import com.aiary.aiary.domain.diary.exception.DiaryNotFoundException;
 import com.aiary.aiary.domain.diary.repository.DiaryRepository;
 import com.aiary.aiary.domain.user.entity.User;
+import com.aiary.aiary.domain.user.entity.UserDetail;
+import com.aiary.aiary.domain.user.exception.UnAuthorizedAccessException;
+import com.aiary.aiary.domain.user.exception.UserNotFoundException;
+import com.aiary.aiary.domain.user.repository.UserRepository;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,11 +18,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
 public class DiaryService {
-
+    private final UserRepository userRepository;
     private final DiaryRepository diaryRepository;
     private final DiaryMapper diaryMapper;
 
@@ -28,8 +33,13 @@ public class DiaryService {
     }
   
     @Transactional
-    public void deleteDiary(Long diaryId){
-        Diary deleteDiary = findDiaryById(diaryId);
+    public void deleteDiary(UserDetail userDetail,  Long diaryId){
+        Diary deleteDiary = diaryRepository.findDiaryWithUser(diaryId).orElseThrow(DiaryNotFoundException::new);
+        User user = userRepository.findUserByEmail(userDetail.getUsername()).orElseThrow(UserNotFoundException::new);
+
+        if (!Objects.equals(user.getId(), deleteDiary.getUser().getId()))
+            throw new UnAuthorizedAccessException();
+
         deleteDiary.delete();
     }
